# New CPU Usage Retrieval and Process Counting Method