Implements a new method of CPU usage retrieval and process counting using .NET

- Uses `[diagnostics.process]::GetProcesses()` to get all running processes
  - Iterates through every process, and converts their CPU time to total CPU utilization using the logical processor count and a timespan of the current time and the processes' StartTime, and adds it to a total
    - I'm not exactly sure how this works yet
- Seems to be surprisingly accurate, but hard to test to what degree though due to CPU usage being quite volatile at small timescales
  - Likely accurate enough for this use case
- Might have some more room for optimization, but it is hard to tell as the tracer I use can't seem to decide what is the heaviest part of the process

- Gets the process count using the variable assigned to the output of `[diagnostics.process]::GetProcesses()`

Speed:
- PowerShell 5:
  - 1035ms → 177.2ms
    - 5.8x Faster
- PowerShell 7:
  - 1035ms → 27.18ms
    - 38x Faster

Bottlenecked Speed:
- PowerShell 5:
  - 1086ms → 658.76ms
    - 1.65x Faster
- PowerShell 7:
  - 1073ms → 121.75ms
    - 8.8x Faster