@rashil2000 2022-11-02
I'm seeing very significant performance regression on cold cache (5.5s vs 2.1s). Most users run winfetch once (during shell startup, or once during a demo etc.), so the cache will always remain cold. 

On warmer cache (repeated runs), there is improvement (1.0s vs 1.6s).