@Carterpersall 2022-11-02
Might be best to go back to the drawing board on this implementation then. There may not be a way to avoid the WMI call here, but there might be a way to improve speed in other places.