@Carterpersall 2022-11-03
I'm currently working on quite a cool solution that gets all running processes and adds together their collective CPU usage. My testing so far has shown that the returned value is accurate while only taking ~80ms